Added autoRepeat-option to Winval
Actually pretty useless in the current context as all deactivations of keys are immediately reverted if it was an autorepeat-keystroke. The users only use isKeyDown, and will never read this between deactivation and immediate reactivation. This change is therefore currently invisible, but maybe the future will hold some use cases?